Fix all fakematches in event object movement
Min changes needed. This also required adjusting faraway_mew's return type.

+++ include/faraway_island.h
@@ -1,7 +1,7 @@
 #ifndef GUARD_FARAWAY_ISLAND_H
 #define GUARD_FARAWAY_ISLAND_H
 
-u32 GetMewMoveDirection(void);
+u8 GetMewMoveDirection(void);
 bool8 ShouldMewShakeGrass(struct ObjectEvent *objectEvent);
 void UpdateFarawayIslandStepCounter(void);
 bool8 ObjectEventIsFarawayIslandMew(struct ObjectEvent *objectEvent);
